Extract services to interfaces
