+++ src/main/java/bot/externalservice/apium/impl/ApiUmServiceImpl.java
@@ -1,0 +1,51 @@
+package bot.externalservice.apium.impl;
+
+import bot.externalservice.apium.ApiUmService;
+import bot.externalservice.apium.configuration.ApiUmConfiguration;
+import bot.externalservice.apium.dto.ApiUmResponseDtoWrapper;
+import bot.externalservice.apium.response.ApiUmResponse;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
+import org.springframework.web.util.UriComponentsBuilder;
+
+@Service
+@Slf4j
+public class ApiUmServiceImpl implements ApiUmService {
+
+    private static final String SERVICE_URL = "https://api.um.warszawa.pl/api/action/dbtimetable_get/?id=e923fa0e-d96c-43f9-ae6e-60518c9f3238";
+
+    private final RestTemplate restTemplate;
+
+    private ApiUmConfiguration apiUmConfiguration;
+
+    @Autowired
+    public ApiUmServiceImpl(ApiUmConfiguration apiUmConfiguration, RestTemplate restTemplate) {
+        this.apiUmConfiguration = apiUmConfiguration;
+        this.restTemplate = restTemplate;
+    }
+
+    @Override
+    public ApiUmResponse getDepartureDetails(String stationId, String platformNumber, String line) {
+        String uri = UriComponentsBuilder.fromHttpUrl(SERVICE_URL)
+                .queryParam("busstopId", stationId)
+                .queryParam("busstopNr", platformNumber)
+                .queryParam("line", line)
+                .queryParam("apikey", apiUmConfiguration.getKey())
+                .toUriString();
+
+        log.info("Calling API UM at " + uri + " for departure details");
+        ApiUmResponseDtoWrapper apiUmResponseDtoWrapper = restTemplate.getForObject(uri, ApiUmResponseDtoWrapper.class);
+
+        ApiUmResponse response = ApiUmResponse.failed();
+        if (apiUmResponseDtoWrapper != null) {
+            if (apiUmResponseDtoWrapper.isSuccess()) {
+                response = new ApiUmResponse(apiUmResponseDtoWrapper, line);
+            } else {
+                log.error("API UM responded with error. Reason: " + apiUmResponseDtoWrapper.getMessage());
+            }
+        }
+        return response;
+    }
+}
