+++ src/test/resources/apium/unknownError.json
@@ -1,0 +1,1 @@
+Error: The table 'aktywnosc' is full
